requirements.txt
# Milvus and Vector Database
pymilvus==2.4.0

# Machine Learning & AI
torch>=1.13.0
torchvision>=0.14.0
transformers>=4.20.0
sentence-transformers>=2.2.0
numpy>=1.21.0
scikit-learn>=1.1.0
opencv-python>=4.6.0
Pillow>=9.2.0

# Data Processing
pandas>=1.4.0
datasets>=2.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Web Framework & API
fastapi>=0.85.0
uvicorn>=0.18.0
streamlit>=1.15.0
gradio>=3.4.0

# Database & Storage
sqlalchemy>=1.4.0
redis>=4.3.0

# Monitoring & Logging
prometheus-client>=0.14.0
grafana-api>=1.0.3

# Development & Testing
jupyter>=1.0.0
notebook>=6.4.0
pytest>=7.1.0
pytest-asyncio>=0.19.0
black>=22.0.0
flake8>=5.0.0

# Utilities
python-dotenv>=0.20.0
tqdm>=4.64.0
requests>=2.28.0
click>=8.1.0
pydantic>=1.10.0

# Text Processing
nltk>=3.7.0
spacy>=3.4.0
textblob>=0.17.0

# Image Processing
albumentations>=1.3.0

# Vector Operations
faiss-cpu>=1.7.0
annoy>=1.17.0

# Async & Concurrency
aiofiles>=0.8.0
asyncio-mqtt>=0.11.0

# Configuration & Environment
pyyaml>=6.0.0
toml>=0.10.0

marshmallow>=3.13,<4.0